@@ -25,7 +25,11 @@ class LayoutBuilderUiCacheContext extends RouteNameCacheContext {
    * {@inheritdoc}
    */
   public function getContext() {
-    return 'is_layout_builder_ui.' . (int) (strpos($this->routeMatch->getRouteName(), 'layout_builder.') !== 0);
+    $route_name = $this->routeMatch->getRouteName();
+    if ($route_name && str_starts_with($route_name, 'layout_builder.')) {
+      return 'is_layout_builder_ui.0';
+    }
+    return 'is_layout_builder_ui.1';
   }
 
 }
