@@ -57,10 +57,11 @@ class TitleResolver implements TitleResolverInterface {
       $arguments = $this->argumentResolver->getArguments($request, $callable);
       $route_title = call_user_func_array($callable, $arguments);
     }
-    elseif ($title = $route->getDefault('_title')) {
+    elseif ($route->hasDefault('_title') && strlen($route->getDefault('_title')) > 0) {
+      $title = $route->getDefault('_title');
       $options = [];
-      if ($context = $route->getDefault('_title_context')) {
-        $options['context'] = $context;
+      if ($route->hasDefault('_title_context')) {
+        $options['context'] = $route->getDefault('_title_context');
       }
       $args = [];
       if (($raw_parameters = $request->attributes->get('_raw_variables'))) {
