@@ -285,7 +285,9 @@ class HtmlRenderer implements MainContentRendererInterface {
     }
 
     // Allow hooks to add attachments to $page['#attached'].
-    $this->invokePageAttachmentHooks($page);
+    $this->renderer->executeInRenderContext(new RenderContext(), function () use (&$page) {
+      $this->invokePageAttachmentHooks($page);
+    });
 
     return [$page, $title];
   }
