@@ -683,7 +683,6 @@
     // element monitoring itself.
     if (e.trigger) {
       $(e.target)
-        .prop('disabled', e.value)
         .closest('.js-form-item, .js-form-submit, .js-form-wrapper')
         .toggleClass('form-disabled', e.value)
         .find('select, input, textarea')
@@ -726,7 +725,11 @@
 
   $document.on('state:checked', (e) => {
     if (e.trigger) {
-      $(e.target).prop('checked', e.value).trigger('change');
+      $(e.target)
+        .closest('.js-form-item, .js-form-wrapper')
+        .find('input')
+        .prop('checked', e.value)
+        .trigger('change');
     }
   });
 
