@@ -84,7 +84,11 @@ class LanguageBlock extends BlockBase implements ContainerFactoryPluginInterface
   public function build() {
     $build = [];
     $type = $this->getDerivativeId();
-    $links = $this->languageManager->getLanguageSwitchLinks($type, Url::fromRouteMatch(\Drupal::routeMatch()));
+    $route_match = \Drupal::routeMatch();
+    // If there is no route match, for example when creating blocks on 404 pages
+    // for logged-in users with big_pipe enabled using the front page instead.
+    $url = $route_match->getRouteObject() ? Url::fromRouteMatch($route_match) : Url::fromRoute('<front>');
+    $links = $this->languageManager->getLanguageSwitchLinks($type, $url);
 
     if (isset($links->links)) {
       $build = [
