@@ -14,7 +14,13 @@ class PreviewTest extends UITestBase {
    *
    * @var array
    */
-  public static $testViews = ['test_preview', 'test_preview_error', 'test_pager_full', 'test_mini_pager', 'test_click_sort'];
+  public static $testViews = [
+    'test_preview',
+    'test_preview_error',
+    'test_pager_full',
+    'test_mini_pager',
+    'test_click_sort',
+  ];
 
   /**
    * {@inheritdoc}
@@ -169,6 +175,30 @@ SQL;
     $this->submitForm($edit = [], 'Update preview');
 
     $this->assertSession()->pageTextContains('Unable to preview due to validation errors.');
+  }
+
+  /**
+   * Tests HTML is filtered from the view title when previewing.
+   */
+  public function testPreviewTitle() {
+    // Update the view and change title with html tags.
+    \Drupal::configFactory()->getEditable('views.view.test_preview')
+      ->set('display.default.display_options.title', '<strong>Test preview title</strong>')
+      ->save();
+
+    $this->drupalGet('admin/structure/views/view/test_preview/edit');
+    $this->assertSession()->statusCodeEquals(200);
+    $this->submitForm([], 'Update preview');
+    $this->assertSession()->pageTextContains('Test preview title');
+    // Ensure allowed HTML tags are still displayed.
+    $this->assertCount(2, $this->xpath('//div[@id="views-live-preview"]//strong[text()=:text]', [':text' => 'Test preview title']));
+
+    // Ensure other tags are filtered.
+    \Drupal::configFactory()->getEditable('views.view.test_preview')
+      ->set('display.default.display_options.title', '<b>Test preview title</b>')
+      ->save();
+    $this->submitForm([], 'Update preview');
+    $this->assertCount(0, $this->xpath('//div[@id="views-live-preview"]//b[text()=:text]', [':text' => 'Test preview title']));
   }
 
 }
