@@ -184,10 +184,12 @@ class SelectComplexTest extends DatabaseTestBase {
   public function testDistinct() {
     $query = $this->connection->select('test_task');
     $query->addField('test_task', 'task');
+    $query->orderBy('task');
     $query->distinct();
-    $query_result = $query->countQuery()->execute()->fetchField();
+    $query_result = $query->execute()->fetchAll(\PDO::FETCH_COLUMN);
 
-    $this->assertEquals(6, $query_result, 'Returned the correct number of rows.');
+    $expected_result = ['code', 'eat', 'found new band', 'perform at superbowl', 'sing', 'sleep'];
+    $this->assertEquals($query_result, $expected_result, 'Returned the correct result.');
   }
 
   /**
