@@ -157,7 +157,7 @@ class KeyValueEntityStorageTest extends UnitTestCase {
 
     $this->moduleHandler->expects($this->exactly(2))
       ->method('invokeAll')
-      ->withConsecutive(['test_entity_type_create'], ['entity_create']);
+      ->willReturnOnConsecutiveCalls(['test_entity_type_create'], ['entity_create']);
     $this->uuidService->expects($this->never())
       ->method('generate');
 
@@ -180,7 +180,7 @@ class KeyValueEntityStorageTest extends UnitTestCase {
 
     $this->moduleHandler->expects($this->exactly(2))
       ->method('invokeAll')
-      ->withConsecutive(['test_entity_type_create'], ['entity_create']);
+      ->willReturnOnConsecutiveCalls(['test_entity_type_create'], ['entity_create']);
     $this->uuidService->expects($this->never())
       ->method('generate');
 
@@ -205,7 +205,7 @@ class KeyValueEntityStorageTest extends UnitTestCase {
 
     $this->moduleHandler->expects($this->exactly(2))
       ->method('invokeAll')
-      ->withConsecutive(['test_entity_type_create'], ['entity_create']);
+      ->willReturnOnConsecutiveCalls(['test_entity_type_create'], ['entity_create']);
     $this->uuidService->expects($this->once())
       ->method('generate')
       ->willReturn('bar');
@@ -247,7 +247,7 @@ class KeyValueEntityStorageTest extends UnitTestCase {
 
     $this->moduleHandler->expects($this->exactly(4))
       ->method('invokeAll')
-      ->withConsecutive(
+      ->willReturnOnConsecutiveCalls(
         ['test_entity_type_presave'],
         ['entity_presave'],
         ['test_entity_type_insert'],
@@ -291,7 +291,7 @@ class KeyValueEntityStorageTest extends UnitTestCase {
       ->method('delete');
     $this->moduleHandler->expects($this->exactly(4))
       ->method('invokeAll')
-      ->withConsecutive(
+      ->willReturnOnConsecutiveCalls(
         ['test_entity_type_presave'],
         ['entity_presave'],
         ['test_entity_type_update'],
@@ -569,7 +569,7 @@ class KeyValueEntityStorageTest extends UnitTestCase {
 
     $this->moduleHandler->expects($this->exactly(8))
       ->method('invokeAll')
-      ->withConsecutive(
+      ->willReturnOnConsecutiveCalls(
         ['test_entity_type_predelete'],
         ['entity_predelete'],
       );
